+++ src/setup/errors.py
@@ -24,3 +24,4 @@ class DiskSpaceError(Exception):
 
     pass
 
+